% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullFRAP_boundary_functions.R
\name{circle_boundary_full_fit}
\alias{circle_boundary_full_fit}
\title{Full-circle FRAP function, diffusion with boundary}
\usage{
circle_boundary_full_fit(t)
}
\arguments{
\item{t}{Time points}

\item{d}{Diffusion coefficient}

\item{h}{Permeability}

\item{radius}{Radius of the bleached circle}
}
\value{
FRAP curve
}
\description{
Full-circle FRAP function, diffusion with boundary
}
