% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullFRAP_no_boundary_functions.R
\name{circle_no_boundary_full_d_fit}
\alias{circle_no_boundary_full_d_fit}
\title{Full-circle FRAP function, diffusion without boundary}
\usage{
circle_no_boundary_full_d_fit(t)
}
\arguments{
\item{t}{Time points}

\item{d}{Diffusion coefficient}

\item{rc}{Radius of the bleached circle}

\item{rl}{Radius of the entire reservoir}
}
\value{
FRAP curve
}
\description{
Full-circle FRAP function, diffusion without boundary
}
