% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frappe.R
\name{frappe}
\alias{frappe}
\title{Frappe}
\usage{
frappe(
  data,
  labels,
  type = c("bar", "line", "pie", "donut", "percentage"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Basic frappe chart
}
