% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frappe2.R
\name{frappe2}
\alias{frappe2}
\title{<Add Title>}
\usage{
frappe2(
  data1,
  data2,
  labels,
  type1 = c("bar", "line"),
  type2 = c("bar", "line"),
  ...,
  heat,
  shade,
  spline,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
