% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{hypergeo}
\alias{hypergeo}
\title{Hypergeometric series}
\usage{
hypergeo(m, z, cutoff = 400)
}
\arguments{
\item{m}{Parameter}

\item{z}{Primary argument}

\item{cutoff}{Cutoff value to prevent rounding errors}
}
\value{
Hypergeometric series
}
\description{
Hypergeometric series
}
