% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfraps.R
\name{mixfraps}
\alias{mixfraps}
\title{mixfraps}
\usage{
mixfraps(
  data1,
  data2,
  data3,
  labels,
  type1 = c("bar", "line"),
  type2 = c("bar", "line"),
  type3 = c("bar", "line"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Mixed frap charts
}
