% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfFRAP_boundary_functions.R
\name{nonbleached_boundary_half_fit}
\alias{nonbleached_boundary_half_fit}
\title{Half-circle FRAP function, diffusion with boundary, non-bleached half}
\usage{
nonbleached_boundary_half_fit(t)
}
\arguments{
\item{t}{Time points}

\item{d}{Diffusion coefficient}

\item{h}{Permeability}

\item{radius}{Radius of the bleached circle}
}
\value{
FRAP curve
}
\description{
Half-circle FRAP function, diffusion with boundary, non-bleached half
}
