% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfFRAP_no_boundary_functions.R
\name{nonbleached_no_boundary_half_rd_fit}
\alias{nonbleached_no_boundary_half_rd_fit}
\title{Half-circle FRAP function, reaction-diffusion without boundary, non-bleached half}
\usage{
nonbleached_no_boundary_half_rd_fit(t)
}
\arguments{
\item{t}{Time points}

\item{d}{Diffusion coefficient}

\item{rc}{Radius of the bleached circle}

\item{rl}{Radius of the entire reservoir}

\item{kon}{Pseudo-association rate}

\item{koff}{Dissociation rate}
}
\value{
FRAP curve
}
\description{
Half-circle FRAP function, reaction-diffusion without boundary, non-bleached half
}
