% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullFRAP_no_boundary_functions.R
\name{reaction}
\alias{reaction}
\title{Full-circle FRAP function, reaction-dominant}
\usage{
reaction(t)
}
\arguments{
\item{t}{Time points}

\item{kon}{Pseudo-association rate}

\item{koff}{Dissociation rate}
}
\value{
FRAP curve
}
\description{
Full-circle FRAP function, reaction-dominant
}
