% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{secant}
\alias{secant}
\title{Secant function used by "zeros"-function to find roots of a*J_{n-1}(a*r) + (h*r-n)*J_{n}(a*r)/r in interval (a0, a1)}
\usage{
secant(r, h, a0, a1, n, prec, tol = 1e-07, niter = 5000)
}
\arguments{
\item{r}{Radius}

\item{h}{Permeability}

\item{a0}{Start of search interval}

\item{a1}{End of search interval}

\item{n}{Order parameter for Bessel functions}

\item{prec}{Precision}

\item{tol}{Tolerance}

\item{niter}{Number of iterations}
}
\value{
Secant
}
\description{
Secant function used by "zeros"-function to find roots of a*J_{n-1}(a*r) + (h*r-n)*J_{n}(a*r)/r in interval (a0, a1)
}
