% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullFRAP_no_boundary_functions.R
\name{soumpasis}
\alias{soumpasis}
\title{Full-circle FRAP function, diffusion}
\usage{
soumpasis(t)
}
\arguments{
\item{t}{Time points}

\item{d}{Diffusion coefficient}

\item{radius}{Radius of the bleached circle}
}
\value{
FRAP curve
}
\description{
Full-circle FRAP function, diffusion
}
