% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{zeros}
\alias{zeros}
\title{function to find roots of J'_{n}(a*r) + h*J_{n}(a*r) = a/2*J_{n-1}(a*r) - a/2*J_{n+1}(a*r) + h*J_{n}(a*r) = J'_{n}(a*r) + h*J_{n}(a*r) = a*J_{n-1}(a*r) + (h*r-n)*J_{n}(a*r)/r in interval (0,m)}
\usage{
zeros(r, h, n, m, prec)
}
\arguments{
\item{r}{Radius}

\item{h}{Permeability}

\item{n}{Order parameter for Bessel functions}

\item{m}{End of interval}

\item{prec}{Precision}
}
\value{
Approximated zeros
}
\description{
function to find roots of J'_{n}(a*r) + h*J_{n}(a*r) = a/2*J_{n-1}(a*r) - a/2*J_{n+1}(a*r) + h*J_{n}(a*r) = J'_{n}(a*r) + h*J_{n}(a*r) = a*J_{n-1}(a*r) + (h*r-n)*J_{n}(a*r)/r in interval (0,m)
}
